refactor: get all md files
